% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choices2options}
\alias{choices2options}
\title{Helper function to turn stack select option}
\usage{
choices2options(x)
}
\arguments{
\item{x}{the choices vector}
}
\description{
Helper function to turn stack select option
}
