% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitAccordion.R
\name{UIkitAccordion}
\alias{UIkitAccordion}
\title{Create a UIkit accordion}
\usage{
UIkitAccordion(..., multiple = FALSE)
}
\arguments{
\item{...}{slot for UIkitAccordionItem.}

\item{multiple}{Whether to enable multiple collapsible events.}
}
\description{
Build an UIkit accordion which can contain UIkitAccordionItem.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitContainer(
    size = "large",
    UIkitGrid(
     child_width = "1-2",
     UIkitAccordion(
     UIkitAccordionItem(
      title = "Item 1",
      opened = TRUE, 
      "Lorem ipsum dolor sit amet, consectetur 
      adipiscing elit, sed do eiusmod tempor 
      incididunt ut labore et dolore magna aliqua"
     ),
     UIkitAccordionItem(
      title = "Item 2",
      "Lorem ipsum dolor sit amet, consectetur 
      adipiscing elit, sed do eiusmod tempor 
      incididunt ut labore et dolore magna aliqua"
     ),
     UIkitAccordionItem(
      title = "Item 3",
      "Lorem ipsum dolor sit amet, consectetur 
      adipiscing elit, sed do eiusmod tempor 
      incididunt ut labore et dolore magna aliqua"
     )
    )
   ) 
  )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
