% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitAccordion.R
\name{UIkitAccordionItem}
\alias{UIkitAccordionItem}
\title{Create a UIkit accordion item}
\usage{
UIkitAccordionItem(..., title = NULL, opened = FALSE)
}
\arguments{
\item{...}{Item content.}

\item{title}{Item title.}

\item{opened}{Whether to start with the item opened. FALSE by default.}
}
\description{
Build an UIkit accordion item which can be embedded in an UIkitAccordion.
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
