% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitArticle.R
\name{UIkitArticle}
\alias{UIkitArticle}
\title{Create an UIkit article}
\usage{
UIkitArticle(..., title = NULL, metadata = NULL, url = NULL)
}
\arguments{
\item{...}{Article content.}

\item{title}{Article title.}

\item{metadata}{Article meta datas.}

\item{url}{Url to give more information.}
}
\description{
Build a container to write articles.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    child_width = "1-2",
    UIkitArticle(
      title = "Heading",
      metadata = "Written by Super User on 12 April 2012. Posted in Blog",
      url = "http://www.google.com",
      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
      sed do eiusmod tempor incididunt ut labore et dolore magna 
      aliqua. Ut enim ad minim veniam, quis nostrud 
      exercitation ullamco laboris nisi ut aliquip"
     )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
