% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitBadge.R
\name{UIkitBadge}
\alias{UIkitBadge}
\title{Create a UIkit badge}
\usage{
UIkitBadge(content)
}
\arguments{
\item{content}{Badge Text. Avoid large text.}
}
\description{
Build an UIkit badge.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitBadge(content = 100)
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
