% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitClose.R
\name{UIkitClose}
\alias{UIkitClose}
\title{Create an UIkit close icon}
\usage{
UIkitClose(size = NULL)
}
\arguments{
\item{size}{Icon size: NULL or "large".}
}
\description{
To close elements
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitCard(
    title = "My card",
    badge = "test",
    hover = TRUE,
    horizontal = TRUE,
    UIkitCardMedia(
      src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
      horizontal = TRUE, 
      position = "left"
    ),
    body = "This is the body"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
