% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitColumns.R
\name{UIkitColumns}
\alias{UIkitColumns}
\title{Create an UIkit column}
\usage{
UIkitColumns(..., n, divider = FALSE)
}
\arguments{
\item{...}{Any UI element. You can also insert inline element using UIkitColumnSpan().}

\item{n}{Number of columns: from 2 to 6.}

\item{divider}{Whether to display a divider between columns. FALSE by default.}
}
\description{
Build a set of columns.
}
\note{
UI content needs to be wrapped in a p() tag.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitColumns(
    n = 2,
    divider = TRUE,
    p("Lorem ipsum dolor sit amet, consectetur 
    adipiscing elit, sed do eiusmod tempor 
    incididunt ut labore et dolore magna aliqua. 
    Ut enim ad minim veniam, quis nostrud 
    exercitation ullamco laboris."
    ),
    p("Duis aute irure dolor in reprehenderit 
    in voluptate velit esse cillum dolore eu 
    fugiat nulla pariatur. Excepteur sint occaecat 
    cupidatat non proident, sunt in culpa qui officia 
    deserunt mollit anim id est laborum."
    ),
    UIkitColumnSpan(
    "All we have to decide is what to do 
    with the time that is given us"
    ),
    p("Ut enim ad minim veniam, quis nostrud 
    exercitation ullamco laboris nisi ut 
    aliquip ex ea commodo consequat. 
    Duis aute irure dolor in reprehenderit 
    in voluptate velit esse cillum dolore 
    eu fugiat nulla pariatur. Excepteur sint
    occaecat cupidatat non proident. Duis aute 
    irure dolor in reprehenderit in voluptate 
    velit esse cillum dolore.")
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
