% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitColumns.R
\name{UIkitColumnSpan}
\alias{UIkitColumnSpan}
\title{Create a UIkit special p tag}
\usage{
UIkitColumnSpan(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
This p tag is designed for displayed inline element
in a column layout
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
