% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitComment.R
\name{UIkitComment}
\alias{UIkitComment}
\title{Create an UIkit comment}
\usage{
UIkitComment(..., src = NULL, author = NULL, date = NULL, style = NULL)
}
\arguments{
\item{...}{Comment element.}

\item{src}{Comment image path, if any.}

\item{author}{Comment author.}

\item{date}{Comment date.}

\item{style}{Comment style: NULL by default or "primary".}
}
\description{
Build a container to write comments.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    child_width = "1-2",
    UIkitComment(
      src = "https://getuikit.com/docs/images/avatar.jpg",
      author = "Germaine",
      date = "12 days ago",
      "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, 
      sed diam nonumy eirmod tempor invidunt ut labore et dolore 
      magna aliquyam erat, sed diam voluptua. At vero eos et accusam 
      et justo duo dolores et ea rebum. Stet clita kasd gubergren, 
      no sea takimata sanctus est Lorem ipsum dolor sit amet."
     ),
     UIkitComment(
      style = "primary",
      src = "https://getuikit.com/docs/images/avatar.jpg",
      author = "Germaine",
      date = "12 days ago",
      "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, 
      sed diam nonumy eirmod tempor invidunt ut labore et dolore 
      magna aliquyam erat, sed diam voluptua. At vero eos et accusam 
      et justo duo dolores et ea rebum. Stet clita kasd gubergren, 
      no sea takimata sanctus est Lorem ipsum dolor sit amet."
     )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
