% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitContainer.R
\name{UIkitContainer}
\alias{UIkitContainer}
\title{Create an UIkit container}
\usage{
UIkitContainer(..., size = NULL)
}
\arguments{
\item{...}{Any UI element.}

\item{size}{Container size: "small", "large" or "expand".}
}
\description{
Build a container to align and center page content.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitContainer(
    UIkitCard(
     title = "My card",
     badge = "test",
     hover = TRUE,
     horizontal = TRUE,
     UIkitCardMedia(
      src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
      horizontal = TRUE, 
      position = "left"
     ),
     body = "This is the body"
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
