% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitCover.R
\name{UIkitCover}
\alias{UIkitCover}
\title{Create an UIkit cover}
\usage{
UIkitCover(src = NULL, type = "img", height = "medium")
}
\arguments{
\item{src}{Media source or path.}

\item{type}{Media type: "image", "video" or "iframe".}

\item{height}{Cover height: "medium" by default but also "small" or "large".}
}
\description{
Expand images, videos or iframes to cover their entire 
container and place your own content on top.
}
\note{
Prefer iframe movie quality than the video one.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitCover(
     src = "https://getuikit.com/docs/images/dark.jpg",
     type = "image"
   ),
   UIkitCover(
     src = "https://quirksmode.org/html5/videos/big_buck_bunny.mp4",
     type = "video"
   ),
   UIkitCover(
     src = "https://www.youtube-nocookie.com/embed/YE7VzlLtp-4",
     type = "iframe"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
