% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitDescriptionList.R
\name{UIkitDescriptionList}
\alias{UIkitDescriptionList}
\title{Create an UIkit description list}
\usage{
UIkitDescriptionList(..., divider = FALSE)
}
\arguments{
\item{...}{Slot dor UIkitDescriptionListItem.}

\item{divider}{Whether to display a vertical divider between items. FALSE by default.}
}
\description{
Easily create nice looking description lists, which come in different styles
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    child_width = "1-2",
    UIkitDescriptionList(
    divider = TRUE,
    UIkitDescriptionListItem(
     title = "Description term",
     "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
    ),
    UIkitDescriptionListItem(
     title = "Description term",
     "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
    ),
    UIkitDescriptionListItem(
     title = "Description term",
     "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
    )
   )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
