% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitDescriptionList.R
\name{UIkitDescriptionListItem}
\alias{UIkitDescriptionListItem}
\title{Create an UIkit description list item}
\usage{
UIkitDescriptionListItem(..., title)
}
\arguments{
\item{...}{Item content.}

\item{title}{Item title}
}
\description{
Create an UIkit description list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
