% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitDivider.R
\name{UIkitDivider}
\alias{UIkitDivider}
\title{Create an UIkit divider}
\usage{
UIkitDivider(small = FALSE)
}
\arguments{
\item{small}{Divider size. FALSE by default. If TRUE, the divider class
will be small.}
}
\description{
Create dividers to separate content and apply different styles to them
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   h1("blabla"),
   UIkitDivider(),
   h1("blabla")
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
