% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitDropdown.R
\name{UIkitDropdown}
\alias{UIkitDropdown}
\title{Create an UIkit dropdown}
\usage{
UIkitDropdown(..., dropName = NULL, mode = "click", position = NULL,
  offset = NULL)
}
\arguments{
\item{...}{Any element.}

\item{dropName}{Button name.}

\item{mode}{Dropdown activation mode. "Click" by default. If null, it is 
displayed by hovering over the button trigger.}

\item{position}{Dropdown position: see \url{https://getuikit.com/docs/dropdown#position}.}

\item{offset}{Custom offset between the dropdown container and the toggle. Numeric value.}
}
\description{
Defines different styles for a toggleable dropdown
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitDropdown(
    position = "bottom-justify",
    mode = NULL,
    dropName = "Test me",
    "Lorem ipsum dolor sit amet, consectetur 
    adipiscing elit, sed do eiusmod tempor incididunt."
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
