% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitFlex.R
\name{UIkitFlex}
\alias{UIkitFlex}
\title{Create an UIkit flex}
\usage{
UIkitFlex(..., horiz_align = NULL, vert_align = NULL, direction = "row")
}
\arguments{
\item{...}{Slot for an UIkitFlexElement.}

\item{horiz_align}{Horizontal alignement style: "left", "center", "right", "between"
and "around".}

\item{vert_align}{Vertical alignement style: "stretch", "top", "middle" and
"bottom".}

\item{direction}{Flex direction: either "row" or "column",}
}
\description{
Utilize the power of flexbox to create a wide range of layouts
}
\note{
You can also include UIkitGridElement instead ot UIkitFlexElement.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitFlex(
    horiz_align = "center",
    vert_align = "middle",
    direction = "column",
    UIkitFlexElement(
     UIkitTile(style = "primary", "Primary")
    ),
    UIkitFlexElement(
     UIkitTile(style = "primary", "Primary")
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
