% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitFlex.R
\name{UIkitFlexElement}
\alias{UIkitFlexElement}
\title{Create an UIkit flex element}
\usage{
UIkitFlexElement(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
Element to insert in a flex layout.
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
