% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitGallery.R
\name{UIkitGallery}
\alias{UIkitGallery}
\title{Launch the shinyUIkit Gallery}
\usage{
UIkitGallery()
}
\description{
A gallery of all components available in shinyUIkit.
}
\examples{
if (interactive()) {

 UIkitallery()

}
}
