% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitGrid.R
\name{UIkitGrid}
\alias{UIkitGrid}
\title{Create an UIkit grid}
\usage{
UIkitGrid(..., gutter = NULL, divider = FALSE)
}
\arguments{
\item{...}{Slot for an UIkitGridElement.}

\item{gutter}{Grid gutter (space bewteen elements): "small", "medium", "large" 
or "collapse".}

\item{divider}{Whether to separate grid elements with a vertical divider.
FALSE by default.}
}
\description{
Create a fully responsive, fluid and nestable grid layout
}
\note{
Do not forget to wrap UI element in a UIkitGridElement before putting
them in a grid layout (see example)!
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    gutter = "small",
    UIkitGridElement(
     UIkitCard(
      width = NULL,
      title = "My card",
      badge = "test",
      hover = TRUE,
      horizontal = FALSE,
      UIkitCardMedia(
        src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
        horizontal = TRUE, 
        position = "left"
      ),
      body = "This is the body"
     )
    ),
    UIkitGridElement(
     UIkitCard(
      width = NULL,
      title = "My card",
      badge = "test",
      hover = TRUE,
      horizontal = FALSE,
      UIkitCardMedia(
        src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
        horizontal = TRUE, 
        position = "left"
      ),
      body = "This is the body"
     )
    ),
    UIkitGridElement(
     UIkitCard(
      width = NULL,
      title = "My card",
      badge = "test",
      hover = TRUE,
      horizontal = FALSE,
      UIkitCardMedia(
        src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
        horizontal = TRUE, 
        position = "left"
      ),
      body = "This is the body"
     )
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
