% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitGrid.R
\name{UIkitGridElement}
\alias{UIkitGridElement}
\title{Create an UIkit grid element}
\usage{
UIkitGridElement(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
Element to insert in a grid layout.
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
