% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitHeading.R
\name{UIkitHeading}
\alias{UIkitHeading}
\title{Create an UIkit h1 custom title}
\usage{
UIkitHeading(title, style)
}
\arguments{
\item{title}{Heading content.}

\item{style}{Heading style: "primary", "hero", "divider", "bullet" and "line".}
}
\description{
Define different styles for headings
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    child_width = "1-2",
    UIkitHeading(
    style = "primary",
    title = "Heading Primary"
    ),
    UIkitHeading(
     style = "hero",
     title = "Heading Hero"
    )
   ),
   UIkitGrid(
    child_width = "1-2",
    UIkitHeading(
     style = "divider",
     title = "Heading Divider"
    ),
    UIkitHeading(
     style = "bullet",
     title = "Heading Bullet"
    )
   ),
   UIkitGrid(
    child_width = "1-2",
    UIkitHeading(
     style = "line",
     title = "Heading Line"
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
