% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitIcon.R
\name{UIkitIcon}
\alias{UIkitIcon}
\title{Create an UIkit icon}
\usage{
UIkitIcon(icon, ratio = NULL, button = FALSE)
}
\arguments{
\item{icon}{Icon name. See \url{https://getuikit.com/docs/icon#library} for the full list.}

\item{ratio}{Icon ratio compared to the normal value. 2 means twice bigger than
the base case.}

\item{button}{Whether to wrap the icon in a button.}
}
\description{
Place scalable vector icons anywhere in your content
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitIcon("paint-bucket"),
   UIkitIcon("home", ratio = 2),
   UIkitIcon("twitter", button = TRUE)
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
