% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitLabel.R
\name{UIkitLabel}
\alias{UIkitLabel}
\title{Create an UIkit label element}
\usage{
UIkitLabel(text = NULL, color = NULL)
}
\arguments{
\item{text}{Label Text.}

\item{color}{Label color: "success", "warning", "danger".}
}
\description{
Indicate important notes and highlight parts of your content
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitLabel(
     color = "danger",
     text = "label"
   ),
   UIkitLabel(
     color = "warning",
     text = "label"
   ),
   UIkitLabel(
     color = "success",
     text = "label"
   ),
   UIkitLabel(text = "label")
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
