% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitLightbox.R
\name{UIkitLightbox}
\alias{UIkitLightbox}
\title{Create an UIkit lightbox element}
\usage{
UIkitLightbox(src = NULL, caption = NULL)
}
\arguments{
\item{src}{Image path or url.}

\item{caption}{Image title, description.}
}
\description{
Create a responsive lightbox gallery with images and videos
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitLightbox(
    src = "https://getuikit.com/docs/images/photo.jpg",
    caption = "Hello"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
