% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarChild}
\alias{UIkitNavbarChild}
\title{Create an UIkit navigation bar child element}
\usage{
UIkitNavbarChild(childName = NULL, active = FALSE)
}
\arguments{
\item{childName}{nav item name.}

\item{active}{Whether the item is active or not. FALSE by default.}
}
\description{
To insert in a navigation bar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
