% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarDivider}
\alias{UIkitNavbarDivider}
\title{Create an UIkit navbar divider}
\usage{
UIkitNavbarDivider()
}
\description{
Perfect to insert in a navbar dropdown
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
