% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarDropdown}
\alias{UIkitNavbarDropdown}
\title{Create an UIkit dropdown menu for a navbar}
\usage{
UIkitNavbarDropdown(..., dropdownName)
}
\arguments{
\item{...}{Any element (input, icon, ... logo), UIkitNavbarDivider, UIkitNavbarChild, UIkitNavbarParent.}

\item{dropdownName}{Dropdown name.}
}
\description{
To insert in a navigation bar
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
