% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarHeader}
\alias{UIkitNavbarHeader}
\title{Create an UIkit navbar header}
\usage{
UIkitNavbarHeader(headerName = NULL)
}
\arguments{
\item{headerName}{Header name.}
}
\description{
Perfect to insert in a navbar dropdown to sort dropdown items
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
