% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarItem}
\alias{UIkitNavbarItem}
\title{Create an UIkit navigation bar item}
\usage{
UIkitNavbarItem(...)
}
\arguments{
\item{...}{Any element (input, icon, ... logo).}
}
\description{
To insert in any item in a navigation bar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
