% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbarParent}
\alias{UIkitNavbarParent}
\title{Create an UIkit navigation bar parent element}
\usage{
UIkitNavbarParent(parentName = NULL)
}
\arguments{
\item{parentName}{nav item name.}
}
\description{
To insert in a navigation bar. Can contain several subitems
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
