% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavItem}
\alias{UIkitNavItem}
\title{Create an UIkit navigation bar item}
\usage{
UIkitNavItem(tabName)
}
\arguments{
\item{tabName}{Switcher item name.}

\item{...}{Any UI element.}
}
\description{
Item to insert in an UIkitNavItems
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
