% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavItems}
\alias{UIkitNavItems}
\title{Create an UIkit navigation bar tabSet}
\usage{
UIkitNavItems(..., id, animation = NULL, mode = "switcher")
}
\arguments{
\item{...}{slot for UIkitNavItem.}

\item{id}{unique id.}

\item{animation}{Apply animation to the switcher content. See \url{https://getuikit.com/docs/animation}.}

\item{mode}{NavItems mode: "switcher" or "tabs". "tabs" by default.}
}
\description{
To insert in a navigation bar.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
