% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavMenu}
\alias{UIkitNavMenu}
\title{Create an UIkit navigation bar}
\usage{
UIkitNavMenu(..., position = "left", sidebarTrigger = FALSE,
  sidebarId = NULL, logo = "home")
}
\arguments{
\item{...}{Slot for UIkitNavbarParent, UIkitNavbarChild and UIkitNavbarDropdown.}

\item{position}{Navbar position: "left", "center" or "right".}

\item{sidebarTrigger}{Whether to display the sidebar trigger. FALSE by default.}

\item{sidebarId}{Sidebar id (unique). Share a common id with the sidebar trigger
and the sidebar itself.}

\item{logo}{Sidebar logo.}
}
\description{
To insert in a UIkitNavbar container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
