% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitProgress.R
\name{UIkitProgress}
\alias{UIkitProgress}
\title{Create an UIkit progress bar}
\usage{
UIkitProgress(value, max)
}
\arguments{
\item{value}{Current value.}

\item{max}{Maximum value.}
}
\description{
Defines progress bars that indicate how far along a process is started
}
\examples{
if(interactive()){
 library(shiny)
 shiny::shinyApp(
   ui = UIkitPage(
     title = "My UIkit application",
     sliderInput(
       inputId = "progressinput", 
       label = "Increase the progress bar",
       value = 1, 
       min = 0, 
       max = 10
     ),
     uiOutput("progress")
   ),
   server = function(input, output) {
     output$progress <- renderUI({
       UIkitProgress(
         value = input$progressinput,
         max = 10
       )
     })
   }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
