% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSection.R
\name{UIkitSection}
\alias{UIkitSection}
\title{Create an UIkit section}
\usage{
UIkitSection(..., title = NULL, style = NULL, preserve = TRUE,
  size = NULL)
}
\arguments{
\item{...}{Slot for UIkitSectionItem.}

\item{title}{Section title.}

\item{style}{Section style: "default", "muted", "primary", "secondary".}

\item{preserve}{Whether to preserve the color of UI elements inside the section.
TRUE by default.}

\item{size}{Section size: "xsmall", "small", "large", "xlarge" or NULL (default).}
}
\description{
Create horizontal layout sections with different background colors and styles
}
\examples{
if(interactive()){
 library(shiny)
 shiny::shinyApp(
   ui = UIkitPage(
     title = "My UIkit application",
     UIkitSection(
      size = "small",
      title = "Section Primary with cards",
      style = "primary",
      UIkitGrid(
       child_width = "expand",
       UIkitSectionItem(
        UIkitCard(
         width = NULL,
         height = "medium",
         title = "My card",
         hover = TRUE,
         horizontal = FALSE,
         header = "This is the header",
         body = "This is the body",
         footer = "This is the footer"
        )
       ),
       UIkitSectionItem(
        UIkitCard(
          width = NULL,
          height = "medium",
          title = "My card",
          hover = TRUE,
          horizontal = FALSE,
          header = "This is the header",
          body = "This is the body",
          footer = "This is the footer"
        )
       )
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
