% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSection.R
\name{UIkitSectionItem}
\alias{UIkitSectionItem}
\title{Create an UIkit section item}
\usage{
UIkitSectionItem(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
Wrap any UI element inside
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
