% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSidebar.R
\name{UIkitSidebar}
\alias{UIkitSidebar}
\title{Create an UIkit sidebar}
\usage{
UIkitSidebar(..., id, overlay = TRUE, flip = TRUE)
}
\arguments{
\item{...}{Any UI elements.}

\item{id}{Sidebar id (should be unique).}

\item{overlay}{Whether to blank out the page when the sidebar is open.
TRUE by default.}

\item{flip}{Whether to open the sidebar on the right side. TRUE by default.}
}
\description{
Create an off-canvas sidebar that slides in and out of the page, 
which is perfect for creating mobile navigations
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
