% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSidebar.R
\name{UIkitSidebarLayout}
\alias{UIkitSidebarLayout}
\title{Create an UIkit sidebar layout}
\usage{
UIkitSidebarLayout(sidebarPanel, mainPanel)
}
\arguments{
\item{sidebarPanel}{Slots for the UIkitSidebar.}

\item{mainPanel}{Slot for UIkitSidebarTrigger and the content of
your app.}
}
\description{
Create a sidebarLayout similar to the original sidebarLayout in shiny
}
\examples{
if(interactive()){
 library(shiny)
 
 shiny::shinyApp(
  ui = UIkitPage(
    title = "My UIkit application",
    UIkitSidebarLayout(
     mainPanel = UIkitGrid(
      plotOutput("plot"),
      UIkitSidebarTrigger(id = "mysidebar")
     ), 
     sidebarPanel = UIkitSidebar(
      id = "mysidebar",
      tagList(
       h3("Test Sidebar"),
       "Lorem ipsum dolor sit amet, consectetur adipiscing 
       elit, sed do eiusmod tempor incididunt ut labore et 
       dolore magna aliqua. Ut enim ad minim veniam, quis 
       nostrud exercitation ullamco laboris nisi ut aliquip 
       ex ea commodo consequat.",
       sliderInput("obs", "Number of observations:",
        min = 0, max = 1000, value = 500
      )
     )
    )
   )
  ),
  server = function(input, output) {
   output$plot <- renderPlot({
    hist(rnorm(input$obs))
   })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
