% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSidebar.R
\name{UIkitSidebarTrigger}
\alias{UIkitSidebarTrigger}
\title{Create an UIkit sidebar trigger element}
\usage{
UIkitSidebarTrigger(id)
}
\arguments{
\item{id}{Sidebar id to toggle (should be unique).}
}
\description{
A button to trigger the sidebar
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
