% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSortable.R
\name{UIkitSortable}
\alias{UIkitSortable}
\title{Create an UIkit sortable grid}
\usage{
UIkitSortable(...)
}
\arguments{
\item{...}{Slot for UIkitSortableItem.}
}
\description{
Create sortable grids and lists to rearrange the order of its elements
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitSortable(
   lapply(X = 1:4, FUN = function(i) {
     UIkitSortableItem(
       UIkitCard(
        width = NULL,
        height = "medium",
        title = paste0("My card ", i),
        hover = TRUE,
        horizontal = FALSE,
        body = "This is the body"
       )
     )
   })
  )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
