% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSortable.R
\name{UIkitSortableItem}
\alias{UIkitSortableItem}
\title{Create an UIkit sortable element}
\usage{
UIkitSortableItem(...)
}
\arguments{
\item{...}{Any UI element to be sorted.}
}
\description{
To put inside a sortable grid
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
