% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSpinner.R
\name{UIkitSpinner}
\alias{UIkitSpinner}
\title{Create an UIkit spinner}
\usage{
UIkitSpinner(ratio = NULL)
}
\arguments{
\item{ratio}{Relative spinner size. If 3, the spinner will be 3 times
bigger than in the normal case.}
}
\description{
Create an animated loading spinner
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitSpinner(),
   UIkitSpinner(ratio = 3),
   UIkitSpinner(ratio = 5)
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
