% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSwitcher.R
\name{UIkitSwitcher}
\alias{UIkitSwitcher}
\title{Create an UIkit switcher nav menu}
\usage{
UIkitSwitcher(..., animation = NULL, mode = "switcher")
}
\arguments{
\item{...}{Slot for UIkitSwitcherItem.}

\item{animation}{Apply animation to the switcher content. See \url{https://getuikit.com/docs/animation}.}

\item{mode}{Switcher mode: either nav pills or tabs. "switcher" by default.}
}
\description{
Dynamically transition through different content panes
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitSwitcher(
    animation = "fade",
    mode = "switcher",
    UIkitSwitcherItem(
     switcherItemName = "tab1",
     active = TRUE
    ),
    UIkitSwitcherItem(
     switcherItemName = "tab2",
     disabled = TRUE
    )
   ),
   
   UIkitSwitcherContent(
    UIkitSwitcherContentItem(
     "Test"
    ), 
    UIkitSwitcherContentItem(
     "blabla"
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
