% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSwitcher.R
\name{UIkitSwitcherContent}
\alias{UIkitSwitcherContent}
\title{Create an UIkit switcher content menu}
\usage{
UIkitSwitcherContent(...)
}
\arguments{
\item{...}{Slot for  UIkitSwitcherContentItem.}
}
\description{
Item to insert in an UI kit tabset
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
