% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSwitcher.R
\name{UIkitSwitcherContentItem}
\alias{UIkitSwitcherContentItem}
\title{Create an UIkit switcher content item}
\usage{
UIkitSwitcherContentItem(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
Item to insert in an UI kit tabset
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
