% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSwitcher.R
\name{UIkitSwitcherItem}
\alias{UIkitSwitcherItem}
\title{Create an UIkit switcher nav item}
\usage{
UIkitSwitcherItem(switcherItemName = NULL)
}
\arguments{
\item{switcherItemName}{Switcher item name.}
}
\description{
Item to insert in an UI kit tabset
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
