% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitTabs.R
\name{UIkitTab}
\alias{UIkitTab}
\title{Create an UIkit tabset}
\usage{
UIkitTab(..., align = NULL, vertical = FALSE, bottom = FALSE)
}
\arguments{
\item{...}{Slot for UIkitTabsItem.}

\item{align}{Tab alignement: "center", "right" or NULL (default).}

\item{vertical}{Whether to display the tab in a vertical mode. FALSE by default.}

\item{bottom}{Whether to flip tab items to the bottom. FALSE by default.}
}
\description{
Create a tabbed navigation with different styles
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitTab(
    UIkitTabItem(
     tabItemName = "tab1",
     active = TRUE
    ),
    UIkitTabItem(
     tabItemName = "tab2",
     disabled = TRUE
    ),
    UIkitTabItem(tabItemName = "tab3")
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
