% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitTabs.R
\name{UIkitTabItem}
\alias{UIkitTabItem}
\title{Create an UIkit tab item}
\usage{
UIkitTabItem(..., tabItemName = NULL, active = FALSE, disabled = FALSE)
}
\arguments{
\item{...}{Tab Item content.}

\item{tabItemName}{Tab item name.}

\item{active}{Whether to active the item or not. FALSE by default.}

\item{disabled}{Whether to disable the tab or not. FALSE by default.}
}
\description{
Item to insert in an UI kit tabset
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
