% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItems.R
\name{UIkitTabItems}
\alias{UIkitTabItems}
\title{Create an UIkit tabItems panel}
\usage{
UIkitTabItems(..., id)
}
\arguments{
\item{...}{Any UI element.}

\item{id}{Unique id element, same as the sidebar menu. See \link{UIkitNavItems}.}
}
\description{
Wrap elements inside when you are using a navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
