% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitWidth.R
\name{UIkitWidth}
\alias{UIkitWidth}
\title{Create an UIkit width}
\usage{
UIkitWidth(..., ratio = "1-2")
}
\arguments{
\item{...}{Any UI element such as card.}

\item{ratio}{By default, the width is set to 1 half or the page, namely "1-2".
Could be set to "1-3", ... until "1-6".}
}
\description{
Set up the element width.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitContainer(
    UIkitWidth(
     ratio = "1-2",
      UIkitCard(
       title = "My card",
       badge = "test",
       hover = TRUE,
       horizontal = TRUE,
       UIkitCardMedia(
         src = "https://getuikit.com/docs/images/light.jpg",
         horizontal = TRUE, 
         position = "left"
       ),
       body = "The card width is 1 quarter of the full page width"
      )
     )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
