% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-button.R
\name{ukButton}
\alias{ukButton}
\title{uikit group button for shiny}
\usage{
ukButton(..., style = "default", size = NULL, width = NULL)
}
\arguments{
\item{...}{UI for the button}

\item{style}{character. UI element style}

\item{size}{character. UI element size}

\item{width}{character. UI element width}
}
\description{
uikit group button for shiny
}
