% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-card.R
\name{ukCardBody}
\alias{ukCardBody}
\title{Create a shinyuikit Card Body}
\usage{
ukCardBody(..., style = NULL)
}
\arguments{
\item{...}{}

\item{style}{character. Can be 'default', 'primary' or 'secondary'.}
}
\description{
Build a UIkit Card Body for Shiny
}
