% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-card.R
\name{ukCardMedia}
\alias{ukCardMedia}
\title{Create a shinyuikit Card Media}
\usage{
ukCardMedia(..., position = NULL)
}
\arguments{
\item{...}{}

\item{position}{}
}
\description{
Build a UIkit Card media for Shiny
}
