% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-checkbox.R
\name{ukCheckBoxInput}
\alias{ukCheckBoxInput}
\title{Create a shinyuikit checkbox input}
\usage{
ukCheckBoxInput(inputId, label, choices = NULL, selected = NULL)
}
\arguments{
\item{inputId}{character. The input identifider used to access the value}

\item{label}{character. Input label}

\item{choices}{character. Named or unnamed vector of choices.}

\item{selected}{character. Selected checkbox input}
}
\description{
UIkit checkbox input for Shiny
}
\examples{
\dontrun{
library(shiny)
shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
   ukCheckBoxInput("select", c("Miles per galon" = "mpg", "Rear axle ratio" = "drat")),
   verbatimTextOutput("selected")
  ),
  server = function(input, output) {
    output$selected <- renderText({
      input$select
    })
  }
)
}
}
