% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-column.R
\name{ukColumn}
\alias{ukColumn}
\title{Create a shinyuikit Card Body}
\usage{
ukColumn(..., ncol = 2, divider = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the card body.}

\item{divider}{logical. Add divider.}
}
\description{
Build a UIkit Card Body for Shiny
}
