% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-container.R
\name{ukContainer}
\alias{ukContainer}
\title{Building container in shinyuikit}
\usage{
ukContainer(..., size = NULL)
}
\arguments{
\item{...}{The UI elements to place in the card}

\item{size}{character. size of the container 'small', 'medium' or 'large'.}
}
\description{
UIkit container for shiny
}
\examples{
\dontrun{
library(shiny)
shinyApp(
 ui = ukPage(
  ukContainer(
    h1("Hello", class = "uk-heading")
  )
 ),
 server = function(input, output) {}
)}
}
\author{
Ahmadou H. Dicko
}
