% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-divider.R
\name{ukDivider}
\alias{ukDivider}
\title{UIkit horizontal divider}
\usage{
ukDivider(icon = FALSE, small = FALSE)
}
\arguments{
\item{icon}{add a icon to the divider}

\item{small}{small divider}
}
\description{
UIkit horizontal divider
}
