% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-grid.R
\name{ukGrid}
\alias{ukGrid}
\title{Create a shinyuikit Grid}
\usage{
ukGrid(..., divider = FALSE, gutter = NULL, match_height = FALSE,
  child_width = NULL)
}
\arguments{
\item{...}{}

\item{divider}{logical. Divider modifier that separate grid cells with lines.}

\item{gutter}{character. size of the gutter 'small', 'medium' or 'large', you can use 'collapse' to remove gutter entirely.}

\item{match_height}{logical.}

\item{child_width}{character. grid with child elements evenly split without the need to apply to each list item within the grid.
you can choose 'half' to have half of their parent container, 'third' take up third of parent container, this apply to 'fourth', 'fifth' or 'sixth'.
'auto' divides the grid into equal units based on content size or 'expand' equal units on available space}
}
\description{
Build a shinyuikit Grid
}
