% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-icon.R
\name{ukIcon}
\alias{ukIcon}
\title{UIkit icons for shiny}
\usage{
ukIcon(name, modifier = NULL, image_path = NULL, ratio = NULL)
}
\arguments{
\item{name}{icon name}

\item{modifier}{link, button or image icon}

\item{ratio}{numeric size of icon}
}
\description{
UIkit icons for shiny
}
