% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-navbar.R
\name{ukNavBar}
\alias{ukNavBar}
\title{Create a shinyuikit NavBar}
\usage{
ukNavBar(..., transparent = FALSE, margin = FALSE, position = NULL,
  options = NA)
}
\arguments{
\item{...}{}
}
\description{
Build a shinyuikit NavBar
}
