% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-page.R
\name{ukPage}
\alias{ukPage}
\title{.. content for \description{} (no empty lines) ..}
\usage{
ukPage(..., title = NULL)
}
\arguments{
\item{...}{}

\item{title}{}
}
\description{
.. content for \details{} ..
}
\author{
Ahmadou H. Dicko
}
