% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-radio.R
\name{ukRadioInput}
\alias{ukRadioInput}
\title{Create a shinyuikit radio}
\usage{
ukRadioInput(inputId, label = NULL, choices = NULL, selected = NULL)
}
\arguments{
\item{inputId}{character. The input identifider used to access the value}

\item{label}{character. Radio label}

\item{choices}{character. Named or unnamed vector of choices.}

\item{selected}{character. Selected radio input}
}
\description{
UIkit radio for Shiny
}
\examples{
\dontrun{
library(shiny)

shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
   ukRadioInput("select", "Choose a col", c("Miles per galon" = "mpg", "Rear axle ratio" = "drat"), selected = "mpg"),
   verbatimTextOutput("selected")
  ),
  server = function(input, output) {
    output$selected <- renderText({
      input$select
    })
  }
)
}
}
