% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-search.R
\name{ukSearchInput}
\alias{ukSearchInput}
\title{Create a shinyuikit search input}
\usage{
ukSearchInput(inputId, label = NULL, placeholder = NULL, width = NULL,
  navbar = FALSE, default_style = TRUE, icon = TRUE)
}
\arguments{
\item{inputId}{character. The input identifider used to access the value.}

\item{label}{character. Input label.}

\item{placeholder}{character. Input placeholder.}

\item{width}{character. Input width.}
}
\description{
UIkit search input for shiny
}
\examples{
\dontrun{
library(shiny)
shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
      ukSearchInput("txt", placeholder = "Type here"),
   verbatimTextOutput("default")
  ),
  server = function(input, output) {
    output$default <- renderText({ input$txt })
  }
)
}
}
