% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-section.R
\name{ukSection}
\alias{ukSection}
\title{Create a shinyuikit Section}
\usage{
ukSection(..., style = NULL, size = NULL, preserve_color = FALSE,
  remove_padding = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the card}

\item{style}{character. Can be 'default', 'primary' or 'secondary'.}

\item{size}{character. Can be 'small' or 'large'.}

\item{preserve_color}{logical. Preserve the original color of the content}

\item{remove_padding}{logical. Remove padding}
}
\description{
Build a shinyuikit Section
}
\examples{
\dontrun{
library(shiny)
shinyApp(
ui = uikitPage(
 ukSection(
   ukContainer(
     h1("Section", class = "uk-heading-primary"),
     p(HTML(paste0("A simple container to divide your page into ", strong("section"), ", like the one you're currently reading"))),
     class = "uk-text-center"
   ),
   style = "muted"
 )
), 
server = function(input, output) {}
)
}
}
