% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-select.R
\name{ukSelectInput}
\alias{ukSelectInput}
\title{Create a shinyuikit select input}
\usage{
ukSelectInput(inputId, label, choices = NULL)
}
\arguments{
\item{inputId}{character.The input identifider used to access the value.}

\item{label}{character. Input label.}

\item{choices}{character. Named or unnamed vector of choices.}
}
\description{
UIkit select input form for Shiny
}
\examples{
\dontrun{
library(shiny)

shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
   ukSelectInput("var", "Select",
     choices = c("Miles per galon" = "mpg", "Displacement" = "disp")),
   tableOutput("plot")
  ),
  server = function(input, output) {
    output$plot <- renderTable({
      mtcars[, c("wt", input$var), drop = FALSE]
    }, rownames = TRUE)
  }
)
}
}
