% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-sidebar.R
\name{ukSidebar}
\alias{ukSidebar}
\title{UIkit side bar}
\usage{
ukSidebar(..., sidebar_id, title = NULL)
}
\arguments{
\item{...}{UI element to include in the sidebar}

\item{sidebar_id}{id of the sidebar}

\item{title}{sidebar header}
}
\description{
create a right side bar
}
