% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-sidebar.R
\name{ukSidebarToggle}
\alias{ukSidebarToggle}
\title{UIkit sidebar toggle button}
\usage{
ukSidebarToggle(..., sidebar_id)
}
\arguments{
\item{...}{UI element for the toggle}

\item{sidebar_id}{id of the created sidebar}
}
\description{
.. content for \details{} ..
}
