% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-subnav.R
\name{ukSubnav}
\alias{ukSubnav}
\title{Subnav for ShinyUIkit}
\usage{
ukSubnav(..., navs = NULL, divider = FALSE, pill = FALSE)
}
\arguments{
\item{...}{UI content}

\item{navs}{tabs}
}
\description{
Subnav for ShinyUIkit
}
