% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-tab.R
\name{ukTab}
\alias{ukTab}
\title{Tab for ShinyUIkit}
\usage{
ukTab(..., title = NULL)
}
\arguments{
\item{...}{UI content}

\item{active}{tabs in active state}

\item{disabled}{tabs in disabled state}

\item{href}{link}
}
\description{
Tab for ShinyUIkit
}
