% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-textarea.R
\name{ukTextAreaInput}
\alias{ukTextAreaInput}
\title{Create a shinyuikit text input}
\usage{
ukTextAreaInput(inputId, label, rows = 3L, placeholder = NULL)
}
\arguments{
\item{inputId}{character. The input identifider used to access the value.}

\item{label}{character. Input label.}

\item{placeholder}{character. Input placeholder.}

\item{row}{integer. Number of rows for text canvas.}
}
\description{
UIkit text input for shiny
}
\examples{
\dontrun{
library(shiny)
shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
   ukTextAreaInput("txt", label = "Input text", placeholder = "Type here"),
   verbatimTextOutput("default")
  ),
  server = function(input, output) {
    output$default <- renderText({ input$txt })
  }
)
}
}
