% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-text.R
\name{ukTextInput}
\alias{ukTextInput}
\title{Create a shinyuikit text input}
\usage{
ukTextInput(inputId, label, placeholder, stacked = TRUE)
}
\arguments{
\item{inputId}{character. The input identifider used to access the value.}

\item{label}{character. Input label.}

\item{placeholder}{character. Input placeholder.}
}
\description{
UIkit text input for shiny
}
\examples{
\dontrun{
library(shiny)
shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
   ukTextInput("txt", label = "Input text", placeholder = "Type here"),
   verbatimTextOutput("default")
  ),
  server = function(input, output) {
    output$default <- renderText({ input$txt })
  }
)
}
}
