% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-width.R
\name{ukWidth}
\alias{ukWidth}
\title{UIkit width div for shiny}
\usage{
ukWidth(..., width = "half")
}
\arguments{
\item{...}{UI element to insert}

\item{width}{width of UI}
}
\description{
UIkit width div for shiny
}
