% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitAlert.R
\name{UIkitAlert}
\alias{UIkitAlert}
\title{Create a UIkit alert}
\usage{
UIkitAlert(..., color = NULL, closable = TRUE)
}
\arguments{
\item{...}{Alert text.}

\item{color}{Alert color: "primary", "success", "warning", "danger".}

\item{closable}{Whether to display a close button. TRUE by default.}
}
\description{
Build an UIkit alert.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitAlert(
    color = "primary",
    h3("Primary Notice"),
    "Lorem ipsum dolor sit amet, consectetur 
    adipiscing elit, sed do eiusmod tempor 
    incididunt"
   ),
   UIkitAlert(
    color = "danger",
    h3("Danger Notice"),
    "Lorem ipsum dolor sit amet, consectetur 
    adipiscing elit, sed do eiusmod tempor 
    incididunt"
   ),
   UIkitAlert(
    color = "warning",
    h3("Warning Notice"),
    "Lorem ipsum dolor sit amet, consectetur 
    adipiscing elit, sed do eiusmod tempor 
    incididunt"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
