% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitCard.R
\name{UIkitCard}
\alias{UIkitCard}
\title{Create an UIkit card}
\usage{
UIkitCard(..., title = NULL, size = NULL, hover = "FALSE",
  style = "default", header = NULL, body = NULL, footer = NULL,
  badge = NULL, horizontal = FALSE, width = "1-2", height = NULL,
  shadow = NULL, shadow_position = NULL, animation = NULL)
}
\arguments{
\item{...}{any element.}

\item{title}{Card title.}

\item{size}{Card size: either NULL, "small" or "large".}

\item{hover}{Whether to trigger a hover effect. FALSE by default.}

\item{style}{Card style: "default", "primary" or "secondary".}

\item{header}{Header content, if any.}

\item{body}{Body content, if any.}

\item{footer}{Footer content, if any.}

\item{badge}{Badge content, if any.}

\item{horizontal}{Whether to display horizontal content. FALSE by default.
If TRUE, use UIkitCardMedia with horizontal set to TRUE!}

\item{width}{Card width: if "1-2", the card will take half of the page,
"1-3" will be one third... If you wrap card in a grid layout, set width to NULL
since the grid will handle it automatically. The sum of all element containing
width should be 1. You can also specify fixed width such as "small", "medium", "large",
"xlarge", "xxlarge".}

\item{height}{Card height: "small", "medium" or "large".}

\item{shadow}{Create a shadow around the box, similar to material design z-depth.
"small", "medium", "large" or "xlarge". NULL by default.}

\item{shadow_position}{Shadow position: NULL or "bottom".}

\item{animation}{NULL by default: card animation if any, see \url{https://getuikit.com/docs/animation#usage}.}
}
\description{
Create layout boxes with different styles
}
\note{
Never use footer and header when horizontal is TRUE!
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitGrid(
    child_width = "expand",
    gutter = "small",
    UIkitGridElement(
     UIkitCard(
      width = NULL,
      height = "medium",
      title = "My card",
      hover = TRUE,
      style = "primary",
      horizontal = FALSE,
      header = "This is the header",
      body = "This is the body",
      footer = "This is the footer"
     )
    ),
    UIkitGridElement(
     UIkitCard(
      width = NULL,
      title = "My card",
      badge = "test",
      horizontal = FALSE,
      UIkitCardMedia(
        src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
        horizontal = TRUE, 
        position = "left"
      ),
      body = "This is the body"
     )
    ),
    UIkitGridElement(
     UIkitCard(
      animation = "shake",
      width = NULL,
      title = "My card",
      hover = TRUE,
      horizontal = TRUE,
      height = "medium",
      UIkitCardMedia(
        src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
        horizontal = TRUE, 
        position = "left"
      ),
      body = "This is the body"
     )
    )
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
