% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitCard.R
\name{UIkitCard}
\alias{UIkitCard}
\title{Create a UIkit card}
\usage{
UIkitCard(..., title = NULL)
}
\arguments{
\item{...}{any element.}

\item{title}{Card title.}
}
\description{
Build an UIkit card.
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
  ui = UIkitPage(
   title = "My UIkit application",
   UIkitCard(
    title = "My card",
    checkboxInput("somevalue", "Some value", FALSE)
   )
  ),
  server = function(input, output) {}
)
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
