% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitCard.R
\name{UIkitCard}
\alias{UIkitCard}
\title{Create a UIkit card}
\usage{
UIkitCard(..., title = NULL, size = NULL, hover = "FALSE",
  style = "default", header = NULL, body = NULL, footer = NULL,
  badge = NULL, horizontal = FALSE, width = "1-2", shadow = NULL,
  shadow_position = NULL)
}
\arguments{
\item{...}{any element.}

\item{title}{Card title.}

\item{size}{Card size: either NULL, "small" or "large".}

\item{hover}{Whether to trigger a hover effect. FALSE by default.}

\item{style}{Card style: "default", "primary" or "secondary".}

\item{header}{Header content, if any.}

\item{body}{Body content, if any.}

\item{footer}{Footer content, if any.}

\item{badge}{Badge content, if any.}

\item{horizontal}{Whether to display horizontal content. FALSE by default.
If TRUE, use UIkitCardMedia with horizontal set to TRUE!}

\item{width}{Card width: if "1-2", the card will take half of the page width,
"1-3" will be one third...}

\item{shadow}{Create a shadow around the box, similar to material design z-depth.
"small", "medium", "large" or "xlarge". NULL by default.}

\item{shadow_position}{Shadow position: NULL or "bottom".}
}
\description{
Build an UIkit card.
}
\note{
Never use footer and header when horizontal is TRUE!
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitCard(
     title = "My card",
     badge = "test",
     hover = TRUE,
     horizontal = TRUE,
     UIkitCardMedia(
      src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
      horizontal = TRUE, 
      position = "left"
     ),
     body = "This is the body"
   ),
   br(),
   UIkitCard(
     title = "My card",
     badge = "test",
     hover = TRUE,
     horizontal = FALSE,
     header = "This is the header",
     style = "secondary",
     UIkitCardMedia(
      src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg",
      horizontal = TRUE, 
      position = "left"
     ),
     body = "This is the body",
     footer = "This is the footer"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
