% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitCard.R
\name{UIkitCardMedia}
\alias{UIkitCardMedia}
\title{Create a UIkit card}
\usage{
UIkitCardMedia(src = NULL, position = "top", horizontal = FALSE)
}
\arguments{
\item{src}{Media url or path.}

\item{position}{Media position: "top" or "bottom". "top" by default. If
horizontal is TRUE, "left" or "right" are also available.}

\item{horizontal}{Media horizontal mode. Only works if the UIkitCard has
horizontal set to TRUE.}
}
\description{
Build an UIkit card.
}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
