% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitGrid.R
\name{UIkitGridElement}
\alias{UIkitGridElement}
\title{Create an UIkit grid element}
\usage{
UIkitGridElement(..., width = NULL)
}
\arguments{
\item{...}{Any UI element.}

\item{width}{Item width. See \url{https://getuikit.com/docs/width}.}
}
\description{
Element to insert in a grid layout.
}
\note{
If the arguement child_width from UIkitGrid is not NULL, you do not
need to provide any individual width argument to the grid element!
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
