% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitImage.R
\name{UIkitImage}
\alias{UIkitImage}
\title{Create an UIkit image}
\usage{
UIkitImage(src = NULL, width, height)
}
\arguments{
\item{src}{Image source or path.}

\item{width}{Image width in pixels.}

\item{height}{Image height in pixels.}
}
\description{
Create a special UIkit ready image tag
}
\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
 ui = UIkitPage(
   title = "My UIkit application",
   UIkitImage(
    src = "https://getuikit.com/docs/images/photo.jpg",
    width = "500",
    height = "500"
   )
 ),
 server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
