% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitNavbar.R
\name{UIkitNavbar}
\alias{UIkitNavbar}
\title{Create an UIkit navigation bar container}
\usage{
UIkitNavbar(..., transparent = FALSE, mode = NULL)
}
\arguments{
\item{...}{Slot for UIkitNavbarnav.}

\item{transparent}{Navbar transparency. FALSE by default.}

\item{mode}{Way to trigger navbar elements: NULL by default (on hover) but also "click".}
}
\description{
Create a navigation bar that can be used for your main site navigation
}
\examples{
if(interactive()){
library(shiny)
shiny::shinyApp(
  ui = UIkitPage(
    title = "My UIkit application",
    
    UIkitSidebarLayout(
      sidebarPanel = UIkitSidebar(
        id = "mysidebar",
        tagList(
          h3("Test Sidebar"),
          "Lorem ipsum dolor sit amet, consectetur adipiscing 
          elit, sed do eiusmod tempor incididunt ut labore et 
          dolore magna aliqua. Ut enim ad minim veniam, quis 
          nostrud exercitation ullamco laboris nisi ut aliquip 
          ex ea commodo consequat.",
          sliderInput("obs", "Number of observations:",
                      min = 0, max = 1000, value = 500
          )
        )
    ),
    mainPanel = UIkitNavbar(
      UIkitNavbarnav(
        sidebarTrigger = TRUE, 
        sidebarId = "mysidebar",
        position = "center",
        UIkitNavbarParent(
          parentName = "Parent"
        ),
        UIkitNavbarChild(
          childName = "Child",
          active = TRUE
        ),
        UIkitNavbarDropdown(
          dropdownName = "Dropdown",
          UIkitNavbarHeader("Group 1"),
          UIkitNavbarChild(
            childName = "item 1"
          ),
          UIkitNavbarChild(
            childName = "item 2"
          ),
          UIkitNavbarDivider(),
          UIkitNavbarHeader("Group 2"),
          UIkitNavbarChild(
            childName = "item 3"
          ),
          UIkitNavbarChild(
            childName = "item 4"
          )
        )
      )
    )
  )
  ),
  server = function(input, output) {}
  )
}

}
\author{
David Granjon, \email{dgranjon@gmail.com}
}
