% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitPage.R
\name{UIkitPage}
\alias{UIkitPage}
\title{Create an UIkit page}
\usage{
UIkitPage(..., title = NULL, navbar = NULL)
}
\arguments{
\item{...}{any element.}

\item{title}{App title.}

\item{navbar}{Navbar content, if any.}
}
\description{
Build an UIkit Shiny app.
}
\examples{
if(interactive()){
library(shiny)
library(shinyUIkit)

shiny::shinyApp(
  ui = UIkitPage(
   title = "My UIkit application",
   sliderInput("obs", "Number of observations:",
    min = 0, max = 1000, value = 500
   ),
   plotOutput("distPlot")
  ),
  server = function(input, output) {
  output$distPlot <- renderPlot({
   hist(rnorm(input$obs))
  })
  }
)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
