% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIkitSwitcher.R
\name{UIkitSwitcherItem}
\alias{UIkitSwitcherItem}
\title{Create an UIkit switcher content item}
\usage{
UIkitSwitcherItem(..., tabName)
}
\arguments{
\item{...}{Any UI element.}

\item{tabName}{Switcher item name.}
}
\description{
Item to insert in an UI kit tabset
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
