% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-button.R
\name{ukButton}
\alias{ukButton}
\alias{ukButtonGroup}
\title{UIkit button for shiny}
\usage{
ukButton(..., style = "default", size = NULL, width = NULL,
  href = FALSE)

ukButtonGroup(...)
}
\arguments{
\item{...}{UI element to insert}

\item{style}{character. UI element style}

\item{size}{character. UI element size}

\item{width}{character. UI element width}

\item{...}{UI for the button group}
}
\description{
UIkit button for shiny

UIkit group button for shiny
}
