% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-card.R
\name{ukCard}
\alias{ukCard}
\alias{ukCardBody}
\alias{ukCardMedia}
\alias{ukCardTitle}
\alias{ukCardHeader}
\alias{ukCardFooter}
\title{Create a shinyuikit Card}
\usage{
ukCard(..., style = NULL, size = NULL, width = NULL, hover = FALSE,
  margin = NULL, padding = NULL, sortable = FALSE)

ukCardBody(..., style = NULL, size = NULL, hover = FALSE, width = NULL,
  height = NULL, margin = NULL, padding = NULL, sortable = FALSE)

ukCardMedia(..., position = "top")

ukCardTitle(title, margin_position = NULL, margin_size = NULL)

ukCardHeader(..., width = NULL)

ukCardFooter(...)
}
\arguments{
\item{...}{The UI elements to place in the card}

\item{style}{character. Can be 'default', 'primary' or 'secondary'.}

\item{size}{character. Can be 'small' or 'large'.}

\item{width}{character. Can be 'default', 'primary' or 'secondary'.}

\item{hover}{logical. Create hover effect on the card.}

\item{sortable}{logical. Allow the card to be sortable.}

\item{height}{character. Can be 'default', 'primary' or 'secondary'.}

\item{position}{character. 'top', 'bottom', 'left' or 'right'}

\item{title}{character. The card title}

\item{...}{The UI elements to place in the card body}

\item{style}{character. Can be 'default', 'primary' or 'secondary'.}

\item{...}{The UI elements to place in the card media}

\item{...}{UI element inside the header}

\item{...}{UI element inside the header}
}
\description{
Build a UIkit Card for Shiny

Build a UIkit Card Body for Shiny

Build a UIkit Card media for Shiny

Build a UIkit Card title for Shiny

Build a UIkit Card Header

Build a UIkit Card Footer
}
\examples{
\dontrun{
library(shiny)
shinyApp(
 ui = ukPage(
   ukContainer(
     ukCard(
       ukCardHeader(ukCardTitle("A nice graph")),
       ukCardBody(plotOutput("graph")),
       ukCardFooter(ukButton(tags$p("More..."), style = "text", href = "#")),
       style = "default")
  )
 ),
 server = function(input, output) {
   output$graph <- renderPlot(hist(rnorm(1000)))
 })


card_line <- function(label)
  tags$li(tags$div(class = "uk-card uk-card-default uk-card-body uk-sortable-handle", label))
ui <- ukPage(
  tags$ul(class = "uk-grid-small uk-child-width-1-3 uk-text-center", `uk-sortable` = "handle: .uk-sortable-handle", `uk-grid` = NA,
        card_line("Item1"),
        card_line("Item2"),
        card_line("Item3"))
)
server <- function(input, output) {}
shinyApp(ui, server)

}

}
