% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-card.R
\name{ukCard}
\alias{ukCard}
\title{Create a shinyuikit Card}
\usage{
ukCard(..., style = NULL, size = NULL, hover = FALSE)
}
\arguments{
\item{...}{}

\item{style}{character. Can be 'default', 'primary' or 'secondary'.}

\item{size}{character. Can be 'small' or 'large'.}

\item{hover}{logical. Create hover effect on the card.}
}
\description{
Build a UIkit Card for Shiny
}
