% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-flex.R
\name{ukFlex}
\alias{ukFlex}
\title{Create a shinyuikit Flex}
\usage{
ukFlex(..., inline = FALSE, wrap = NULL, horizontal_align = NULL,
  vertical_align = NULL, direction = NULL, width = NULL)
}
\arguments{
\item{...}{The UI elements to place in the Flex element}

\item{inline}{logical.}

\item{wrap}{character.}

\item{horizontal_align}{logical.}

\item{vertical_align}{logical.}

\item{direction}{character. Define the axis that flex items are placed and their direction, by default items run horizontaly from left to right as does the 'row' option.}

\item{width}{character. Define Flex item width}
}
\description{
Build a shinyuikit Flex
}
