% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-navbar.R
\name{ukNavbar}
\alias{ukNavbar}
\alias{ukNavbarDropdown}
\alias{ukNavbarItem}
\title{Create a shinyuikit Navbar}
\usage{
ukNavbar(..., transparent = FALSE, margin = FALSE, position = "left",
  overlay = FALSE, options = NA)

ukNavbarDropdown(...)

ukNavbarItem(...)
}
\arguments{
\item{...}{The UI elements to place in the navigation bar}

\item{transparent}{The UI elements to place in the navigation bar}

\item{margin}{The UI elements to place in the navigation bar}

\item{position}{The UI elements to place in the navigation bar}

\item{options}{The UI elements to place in the navigation bar}

\item{...}{The UI elements to place in the navigation bar dropdown}

\item{...}{The UI elements to place in the navigation bar dropdown}
}
\description{
Build a shinyuikit Navbar
}
