% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-navbar.R
\name{ukNavBar}
\alias{ukNavBar}
\alias{ukNavBarDropDown}
\title{Create a shinyuikit NavBar}
\usage{
ukNavBar(..., transparent = FALSE, margin = FALSE, position = NULL,
  options = NA)

ukNavBarDropDown(...)
}
\arguments{
\item{...}{The UI elements to place in the navigation bar}

\item{...}{The UI elements to place in the navigation bar dropdown}
}
\description{
Build a shinyuikit NavBar
}
