% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-page.R
\name{ukPage}
\alias{ukPage}
\title{Create a shinyuikit page}
\usage{
ukPage(..., title = NULL)
}
\arguments{
\item{...}{}

\item{title}{character. The title of the page}
}
\description{
Build a shinyuikit pages
}
\examples{
\dontrun{
shiny::shinyApp(
  ui = ukPage(
   tags$h1("Hello UIkit"),
  ),
  server = function(input, output) {}
)
}
}
