% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uikit-sidebar.R
\name{ukSidebarLayout}
\alias{ukSidebarLayout}
\alias{ukSidebar}
\alias{ukSidebarToggle}
\alias{ukSidebarClose}
\title{UIkit sidebar layout}
\usage{
ukSidebarLayout(...)

ukSidebar(..., sidebar_id, title = NULL)

ukSidebarToggle(..., sidebar_id)

ukSidebarClose()
}
\arguments{
\item{...}{UI element to add the layout}

\item{sidebar_id}{id of the sidebar}

\item{title}{sidebar header}

\item{...}{UI element to include in the sidebar}

\item{...}{UI element for the toggle}

\item{sidebar_id}{id of the created sidebar}
}
\description{
UIkit sidebar layout

create a right side bar

UIkit sidebar toggle button

Button to close the Sidebar
}
